+++ SDLProject/include/Point.hpp
@@ -5,13 +5,13 @@
 class Point
 {
 private:
-	int Point_X;
-	int Point_Y;
+	double Point_X;
+	double Point_Y;
 public:
-	inline Point(int x = 0, int y = 0) : Point_X(x), Point_Y(y) {};
+	inline Point(double x = 0, double y = 0) : Point_X(x), Point_Y(y) {};
 
-	inline int getX() { return Point_X; }
-	inline int getY() { return Point_Y; }
+	inline double getX() { return Point_X; }
+	inline double getY() { return Point_Y; }
 
 	inline double setX(double x) { return Point_X = x; }
 	inline double setY(double y) { return Point_Y = y; }
